backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
python-dotenv
openai
requests
pillow
anthropic
